+++ CHANGELOG.md
@@ -1,6 +1,7 @@
 # svelte-mui changelog
 
 ## unpublish
+- Menu add 'duration' (mS) prop for Popover animation, 300 by default
 - Modal
 - Popover